1.0.38  Build for several Issues
API-186082 Impelement HTTP support
http support has been implemented, which will simplify local testing of modules.
If the protocol is not set, then https is used by default, if the url starts with the protocol (http or https), then it is used

API-217905 WorkPlan.read method doesn't work when workplanId parameter is used

Bug fixed: in the read method of the WorkPlan class, when the value of the workplanId parameter is passed, trackorId is used instead of workplanId

API-175773 Use log level 'Error' by default
Fixed a bug in which, if the logLevelName parameter was missing in the initialization of the IntegrationLog class, an exception appeared. Now by default logLevelName is "Error"